src/lambda_codebase/initial_commit/bootstrap_repository/adf-build/shared/cdk/requirements.txt
boto3
aws_cdk.aws_codepipeline
aws-cdk.aws-codepipeline-actions
aws_cdk.aws_codebuild
aws_cdk.aws_codecommit
aws-cdk.aws-s3
aws-cdk.aws-iam
aws-cdk.aws-kms
